+++ vscode-extension/CHANGELOG.md
@@ -5,6 +5,14 @@ All notable changes to the "Python Import Sorter" extension will be documented i
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.1.1] - 2025-06-10
+
+### Fixed
+- The file being completely emptied after failing to format
+
+### Removed
+- The notification from a single file format
+
 ## [1.1.0] - 2025-06-09
 
 ### Added